% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.gcdnet.R
\name{coef}
\alias{coef}
\title{Extract Model Coefficients}
\usage{
coef(object, ...)
}
\arguments{
\item{object}{an object for which the extraction of model coefficients is
meaningful.}

\item{\ldots}{other arguments.}
}
\value{
Coefficients extracted from the model object \code{object}.
}
\description{
\code{coef} is a generic function which extracts model coefficients from
objects returned by modeling functions. \code{coefficients} is an
\emph{alias} for it.
}
\seealso{
\code{\link{coef.gcdnet}}, \code{\link{coef.erpath}},
  \code{\link{coef.lspath}}, \code{\link{coef.hsvmpath}},
  \code{\link{coef.logitpath}}, \code{\link{coef.sqsvmpath}}.
}
