% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.gcdnet.R
\name{predict}
\alias{predict}
\title{Model predictions}
\usage{
predict(object, ...)
}
\arguments{
\item{object}{a model object for which prediction is desired.}

\item{\dots}{additional arguments affecting the predictions produced.}
}
\value{
The form of the value returned by \code{predict} depends on the
  class of its argument. See the documentation of the particular methods for
  details of what is produced by that method.
}
\description{
\code{predict} is a generic function for predictions from the results of
various model fitting functions. The function invokes particular
\emph{methods} which depend on the \code{\link{class}} of the first
argument.
}
\seealso{
\code{\link{predict.gcdnet}}, \code{\link{predict.erpath}},
  \code{\link{predict.lspath}}, \code{\link{predict.hsvmpath}},
  \code{\link{predict.logitpath}}, \code{\link{predict.sqsvmpath}}.
}
